+++ man/man.md
@@ -9,8 +9,8 @@
 # SYNOPSIS
 
 ```
-falkor-commander [(-- <answers>...)]
-falkor-commander [(-- <answers>...)]
+falkor-commander [(--scope <scope>)] [(--keyword <keyword>)] [(-- <answers>...)]
+falkor-commander [(--s <scope>)] [(--k <keyword>)] [(-- <answers>...)]
 falkor-commander (-v | --version | -h | --help)
 ```
 
@@ -26,8 +26,14 @@ Description
 `-h`, `--help`
 :   Show help and exit
 
+`-s <scope>`, `--scope <scope>`
+:   The scope to look for plugins under **node_modules** (default: **@falkor**)
+
+`-k <keyword>`, `--keyword <keyword>`
+:   The keyword to look for in plugin candidates' **package.json** (default: **@falkor-plugin**)
+
 `-- <answers>...`
-:   Treat all positional arguments after double dash as buffered answers
+:   Treat all positional arguments after double dash as buffered input
 
 # COPYRIGHT
