+++ src/cli/index-cli.ts
@@ -14,13 +14,15 @@ version ${getOwnVersion()}
 Description
 
 Usage:
-  falkor-commander [(-- <answers>...)]
-  falkor-commander [(-- <answers>...)]
+  falkor-commander [(--scope <scope>)] [(--keyword <keyword>)] [(-- <answers>...)]
+  falkor-commander [(--s <scope>)] [(--k <keyword>)] [(-- <answers>...)]
   falkor-commander (-v | --version | -h | --help)
 
 Options:
-  -v, --version    Show version and exit
-  -h, --help       Show this screen and exit
-  -- <answers>...  Treat all positional arguments after double dash as buffered answers
+  -v, --version                      Show version and exit
+  -h, --help                         Show this screen and exit
+  -s <scope>, --scope <scope>        The scope to look for plugins under 'node_modules' [default: @falkor]
+  -k <keyword>, --keyword <keyword>  The keyword to look for in plugin candidates' 'package.json' [default: @falkor-plugin]
+  -- <answers>...                    Treat all positional arguments after double dash as buffered input
 `);
 };
